Feat: 피드 하단 생성정보 컴포넌트 구현 (#176)
